refactor: separate onConnectRequested impl between web and native (#3344)
