+++ packages/thirdweb/src/wallets/coinbase/utils.ts
@@ -1,0 +1,6 @@
+import type { ProviderInterface } from "@coinbase/wallet-sdk";
+
+export async function showCoinbasePopup(provider: ProviderInterface) {
+  // biome-ignore lint/suspicious/noExplicitAny: based on the latest CB SDK - scary but works
+  await (provider as any)?.communicator?.waitForPopupLoaded();
+}
